Move lots of files to follow FSD idea
