+++ src/features/logout/index.ts
@@ -1,0 +1,1 @@
+export { LogoutButton } from './components/LogoutButton/LogoutButton';
