+++ src/shared/icons/RosatomLogo.tsx
