+++ src/shared/icons/index.ts
