+++ src/widgets/CreateDocumentWidget/index.ts
@@ -1,0 +1,1 @@
+export { CreateDocumentWidget } from './components/CreateDocumentWidget/CreateDocumentWidget';
